requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asteroid-filterbanks==0.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.13.3
bibtexparser==2.0.0b8
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-sdr==0.0.2
click==8.1.8
clldutils==3.24.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
ConfigArgParse==1.7
cssselect==1.2.0
csvw==3.5.1
ctc_segmentation==1.7.4
Cython==3.0.11
datasets==3.3.1
debugpy==1.8.12
decorator==5.1.1
dill==0.3.8
Distance==0.1.3
dlinfo==2.0.0
editdistance==0.8.1
einops==0.8.0
espnet==202412
espnet-model-zoo==0.1.7
espnet-tts-frontend==0.0.3
exceptiongroup==1.2.2
executing==2.2.0
fake-useragent==2.0.3
fast-bss-eval==0.1.3
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
g2p-en==2.1.0
h5py==3.12.1
huggingface-hub==0.28.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib-metadata==4.13.0
inflect==7.5.0
ipykernel==6.29.5
ipython==8.32.0
isodate==0.7.2
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaldiio==2.18.0
language-tags==1.2.0
lazy_loader==0.4
librosa==0.9.2
llvmlite==0.43.0
lxml==5.3.0
lxml_html_clean==0.4.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.23.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
parse==1.20.2
parso==0.8.4
pexpect==4.9.0
phonemizer==3.3.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pyee==11.1.1
Pygments==2.19.1
pylatexenc==2.10
pyparsing==3.2.1
pypinyin==0.44.0
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-iso639==2025.1.28
pytz==2025.1
pyworld==0.3.5
PyYAML==6.0.2
pyzmq==26.2.1
rdflib==7.1.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-html==0.10.0
resampy==0.4.3
rfc3986==1.5.0
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
segments==2.2.1
sentencepiece==0.1.97
six==1.17.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
torch-complex==0.4.4
torchaudio==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
triton==3.2.0
typeguard==4.4.1
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==1.26.20
w3lib==2.3.1
wcwidth==0.2.13
websockets==10.4
wikipron==1.3.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
